Rename commissioning headers as service runtime
