+++ openquatt/includes/service/oq_service_logic.h
