+++ openquatt/includes/service/tasks/oq_boiler_task_logic.h
@@ -5,7 +5,7 @@
 #include <stdio.h>
 #include <string>
 
-#include "../oq_commissioning_runtime.h"
+#include "../oq_service_runtime.h"
 
 namespace oq_boiler_task {
