+++ openquatt/includes/service/tasks/oq_flow_autotune_logic.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#include "../oq_commissioning_runtime.h"
+#include "../oq_service_runtime.h"
 
 namespace oq_flow_autotune {
