+++ openquatt/includes/service/tasks/oq_manual_hp_logic.h
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "../../control/oq_thermal_request_logic.h"
-#include "../oq_commissioning_runtime.h"
+#include "../oq_service_runtime.h"
 
 namespace oq_manual_hp {
