+++ openquatt/oq_packages_common.yaml
@@ -6,7 +6,7 @@
 oq_runtime_packages: !include packages/10_runtime.yaml
 oq_strategy_packages: !include packages/20_strategies.yaml
 oq_thermal_pipeline_packages: !include packages/30_thermal_pipeline.yaml
-oq_commissioning_task_packages: !include packages/40_commissioning_tasks.yaml
+oq_service_task_packages: !include packages/40_service_tasks.yaml
 oq_integration_packages: !include packages/50_integrations.yaml
 oq_ui_runtime_packages: !include packages/60_ui_runtime.yaml
 oq_heatpump1_package: !include packages/90_heatpump1.yaml
