+++ openquatt/packages/40_service_tasks.yaml
@@ -1,5 +1,5 @@
 # ==============================================================================
-# OpenQuatt - Commissioning Task Packages
+# OpenQuatt - Service Task Packages
 # ==============================================================================
 packages:
   oq_flow_autotune: !include ../oq_flow_autotune.yaml
